requirements.txt
cv2
numpy
PIL
py-applescript
pyobjc-framework-Quartz